refactor: PageLoader, SocialList, UpdateDate components
