+++ src/components/UpdateDate/index.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react';
 
-import styles from './UpdateDate.module.scss';
+import styles from './styles.module.scss';
 
 type UpdateDateProps = {
   className?: string;
@@ -9,13 +9,11 @@ type UpdateDateProps = {
 
 const DEFAULT_TIME = '00:00 PM';
 
-function UpdateDate({ className, time = DEFAULT_TIME }: UpdateDateProps) {
+export const UpdateDate = memo(({ className, time = DEFAULT_TIME }: UpdateDateProps) => {
   return (
     <p className={`text-light-xl ${styles.updateDate} ${className ?? ''}`}>
       <span className={styles.circle} />
       Last updated at {time}
     </p>
   );
-}
-
-export default memo(UpdateDate);
+});
