+++ src/pages/BankCard/BankCard.tsx
@@ -2,7 +2,7 @@ import { PureComponent } from 'react';
 
 import { ElasticSearch } from '@/components/ElasticSearch';
 import { MyMap } from '@/components/MyMap';
-import UpdateDate from '@/components/UpdateDate';
+import { UpdateDate } from '@/components/UpdateDate';
 import { BANKS_DATA } from '@/constants';
 import { filterBanksByCurrencies, getBanksCurrencies, getTimeFromDate } from '@/utils';
