+++ src/pages/Home/Home.tsx
@@ -4,7 +4,7 @@ import { useSearchParams } from 'react-router-dom';
 import { ConversionForm } from '@/components/ConversionForm';
 import { CurrenciesList } from '@/components/CurrenciesList';
 import { Modal } from '@/components/UI';
-import UpdateDate from '@/components/UpdateDate';
+import { UpdateDate } from '@/components/UpdateDate';
 import { CURRENCY_MODAL_QUERY_KEY } from '@/constants';
 import { useQuotes } from '@/hooks';
