+++ src/pages/Timeline/Timeline.tsx
@@ -3,7 +3,7 @@ import { PureComponent } from 'react';
 import { CandleStick } from '@/components/CandleStick';
 import { HistoryForm } from '@/components/HistoryForm';
 import { Button, Modal } from '@/components/UI';
-import UpdateDate from '@/components/UpdateDate';
+import { UpdateDate } from '@/components/UpdateDate';
 import { DEFAULT_OHLC_PAIRS } from '@/constants';
 import { CandleStickData, OhlcvResponseType } from '@/types';
 import { getTimeFromDate } from '@/utils';
